stake pipeline beta

+++ db/migrate/20211104090031_create_stake_account_histories.rb
@@ -1,15 +1,15 @@
 class CreateStakeAccountHistories < ActiveRecord::Migration[6.1]
   def change
     create_table :stake_account_histories do |t|
-      t.integer :account_balance
+      t.bigint :account_balance
       t.integer :activation_epoch
-      t.integer :active_stake
-      t.integer :credits_observed
-      t.integer :deactivating_stake
+      t.bigint :active_stake
+      t.bigint :credits_observed
+      t.bigint :deactivating_stake
       t.integer :deactivation_epoch
-      t.integer :delegated_stake
+      t.bigint :delegated_stake
       t.string :delegated_vote_account_address
-      t.integer :rent_exempt_reserve
+      t.bigint :rent_exempt_reserve
       t.string :stake_pubkey
       t.string :stake_type
       t.string :staker
